Uploaded complete project
The core skeleton of the operating system was given to me for the
class, and I filled in the various functionality over the course. The
files that aren't c files are for use by the program. Typing 'help' in
the prompt will show a list of available commands. I also implemented
backspacing, inserting, and command-line recall.